+++ NEWS.md
@@ -28,6 +28,8 @@
 
 - Support for creating an Spark extension package using `spark_extension()`.
 
+- Add support for repositories in `spark_dependency()`.
+
 # Sparklyr 1.0.0
 
 ### Arrow